♻️ Move V1 server migration query to own module
Hope is to eventually remove this on-the-fly migration once the database has
actually been migrated

fixup! 🐛 Properly calculate wind direction average
